* commit 'ba9ebaf2d96e0f8726699082b6ab173865c0409e':
  Bugfix: Nope - that didn't work, code functioned but then quantity appeared as 0 for values between greater han 0 but less than one (e.g. items that are sold in weight quantities). Have changed data format from integer to float across board. It now shows everything in qty field with a decimal...but it works!
  Bugfix: AddReceipt.java - receiptLineItem.quantity in some instances can be a float, not an integer. Tweaked to parse as a float but then cast to integer. Need to re-check all six receipts to confirm function.
  Sign up screen - enable spinners to request focus
  Add feature to enable preview of password entered in login & sign up screens
  Update dependencies in build.gradle & .gitignore
  More bug fixes: strings.xml - updated hint text to specify measurements expected for height/weight activity_signup.xml - added imeoptions:actionNext to each editText to enable next button to go to next field.
  Bugfixes based on testing - AddReceipt.java - Woolworths receipts not being handled correctly in some instances - changed parsing logic MainActivity.java - added override method to close application if back button is clicked in MainActivity Signup.java - implement feature to request focus of next EditText element after gender spinner is populated Signup.java, Welcome.java - finish activity after navigated away from this screen to prevent going back to that activity.
  testing more UX fixes - FAQ/about screens
  Pushing some UX fixes - temporary - for issues identified in testing

+++ app/src/main/java/com/stevecrossin/grocerytracker/models/ReceiptLineItem.java
@@ -3,7 +3,7 @@ package com.stevecrossin.grocerytracker.models;
 public class ReceiptLineItem {
     public String itemDescription;
     public float unitPrice;
-    public int quantity;
+    public float quantity;
     public float price;
 
     public ReceiptLineItem() {
@@ -13,7 +13,7 @@ public class ReceiptLineItem {
         String[] columns = line.split(",");
         itemDescription = (columns.length > 0) ? columns[0] : "";
         unitPrice = (columns.length > 1) ? Float.parseFloat(columns[1]) : 0.0f;
-        quantity = (columns.length > 2) ? Integer.parseInt(columns[2]) : 0;
+        quantity = (columns.length > 2) ? Float.parseFloat(columns[2]) : 0.0f;
         price = (columns.length > 3) ? Float.parseFloat(columns[3]) : 0.0f;
     }
